# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HubCredo API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings/outreach:
    get:
      operationId: getOutreachSettings
      tags: [settings]
      summary: Get outreach settings for current user
      responses:
        "200":
          description: Outreach settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachSettings"
        "404":
          description: No settings found

    post:
      operationId: createOutreachSettings
      tags: [settings]
      summary: Create outreach settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutreachSettingsInput"
      responses:
        "201":
          description: Created settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachSettings"

    patch:
      operationId: updateOutreachSettings
      tags: [settings]
      summary: Update outreach settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OutreachSettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OutreachSettings"

  # Stack Recommendations
  /stacks:
    get:
      operationId: listStacks
      tags: [stacks]
      summary: List stack recommendations for current user
      responses:
        "200":
          description: Stack recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StackRecommendation"

    post:
      operationId: createStack
      tags: [stacks]
      summary: Generate and save a stack recommendation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StackInput"
      responses:
        "201":
          description: Created stack recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StackRecommendation"
  /stacks/current:
    get:
      operationId: getCurrentStack
      tags: [stacks]
      summary: Get the current stack recommendation
      responses:
        "200":
          description: Current stack recommendation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StackRecommendation"
        "404":
          description: No stack found

  # Lead Lists
  /lead-lists:
    get:
      operationId: listLeadLists
      tags: [leads]
      summary: List lead lists for current user
      responses:
        "200":
          description: Lead lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeadList"

    post:
      operationId: createLeadList
      tags: [leads]
      summary: Create a new lead list and trigger n8n scraping
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadListInput"
      responses:
        "201":
          description: Created lead list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadList"
  /lead-lists/current:
    get:
      operationId: getCurrentLeadList
      tags: [leads]
      summary: Get the most recent lead list
      responses:
        "200":
          description: Current lead list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadList"
        "404":
          description: No lead list found
  /lead-lists/{id}:
    get:
      operationId: getLeadList
      tags: [leads]
      summary: Get a specific lead list
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lead list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadList"

  # Leads
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads with optional filters
      parameters:
        - name: lead_list_id
          in: query
          schema:
            type: string
        - name: review_status
          in: query
          schema:
            type: string
        - name: industry
          in: query
          schema:
            type: string
        - name: company_size
          in: query
          schema:
            type: string
        - name: hq_country
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
  /leads/{id}/review:
    patch:
      operationId: reviewLead
      tags: [leads]
      summary: Approve or reject a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadReview"
      responses:
        "200":
          description: Updated lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/bulk-review:
    post:
      operationId: bulkReviewLeads
      tags: [leads]
      summary: Bulk approve or reject leads
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkLeadReview"
      responses:
        "200":
          description: Updated leads count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkReviewResult"

  # Dashboard Summary (wow endpoint)
...
  /webhooks/trigger-analysis:
    post:
      operationId: triggerAnalysis
      tags: [webhooks]
      summary: Trigger n8n website analysis webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TriggerAnalysisInput"
      responses:
        "200":
          description: Webhook triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookResult"
...
components:
  schemas:
...
    OutreachSettings:
      type: object
      required: [id, user_id]
      properties:
        id:                       { type: string }
        user_id:                  { type: string }
        email_enabled:            { type: boolean }
        linkedin_enabled:         { type: boolean }
        monthly_lead_target:      { type: ["integer", "null"] }
        monthly_email_target:     { type: ["integer", "null"] }
        monthly_linkedin_target:  { type: ["integer", "null"] }
        messaging_framework:      { type: ["string", "null"] }
        is_current:               { type: boolean }
        created_at:               { type: ["string", "null"] }
        updated_at:               { type: ["string", "null"] }
    OutreachSettingsInput:
      type: object
      properties:
        email_enabled:            { type: boolean }
        linkedin_enabled:         { type: boolean }
        monthly_lead_target:      { type: integer }
        monthly_email_target:     { type: integer }
        monthly_linkedin_target:  { type: integer }
        messaging_framework:      { type: string }
    OutreachSettingsUpdate:
      type: object
      properties:
        email_enabled:            { type: boolean }
        linkedin_enabled:         { type: boolean }
        monthly_lead_target:      { type: integer }
        monthly_email_target:     { type: integer }
        monthly_linkedin_target:  { type: integer }
        messaging_framework:      { type: string }
    ToolItem:
      type: object
      required: [slug, tool_name, category, reason, phase]
      properties:
        slug:       { type: string }
        tool_name:  { type: string }
        category:   { type: string }
        reason:     { type: string }
        is_required: { type: boolean }
        phase:      { type: string }
    StackRecommendation:
      type: object
      required: [id, user_id, tools]
      properties:
        id:          { type: string }
        user_id:     { type: string }
        icp_id:      { type: ["string", "null"] }
        settings_id: { type: ["string", "null"] }
        tools:       { type: array, items: { $ref: "#/components/schemas/ToolItem" } }
        is_current:  { type: boolean }
        created_at:  { type: ["string", "null"] }
    StackInput:
      type: object
      required: [icp_id, settings_id]
      properties:
        icp_id:      { type: string }
        settings_id: { type: string }
...
    LeadList:
      type: object
      required: [id, user_id, status]
      properties:
        id:                 { type: string }
        user_id:            { type: string }
        icp_id:             { type: ["string", "null"] }
        label:              { type: string }
        source:             { type: string }
        total_count:        { type: integer }
        status:             { type: string }
        processing_status:  { type: ["string", "null"] }
        error_message:      { type: ["string", "null"] }
        created_at:         { type: ["string", "null"] }
        updated_at:         { type: ["string", "null"] }
    LeadListInput:
      type: object
      required: [icp_id]
      properties:
        icp_id:  { type: string }
        label:   { type: string }
    Lead:
      type: object
      required: [id, lead_list_id, user_id]
      properties:
        id:                   { type: string }
        lead_list_id:         { type: string }
        user_id:              { type: string }
        first_name:           { type: ["string", "null"] }
        last_name:            { type: ["string", "null"] }
        email:                { type: ["string", "null"] }
        linkedin_url:         { type: ["string", "null"] }
        job_title:            { type: ["string", "null"] }
        seniority:            { type: ["string", "null"] }
        department:           { type: ["string", "null"] }
        company_name:         { type: ["string", "null"] }
        company_domain:       { type: ["string", "null"] }
        company_size:         { type: ["string", "null"] }
        industry:             { type: ["string", "null"] }
        hq_country:           { type: ["string", "null"] }
        hq_city:              { type: ["string", "null"] }
        profile_picture_url:  { type: ["string", "null"] }
        research_blurb:       { type: ["string", "null"] }
        review_status:        { type: string }
        reviewed_at:          { type: ["string", "null"] }
        rejection_reason:     { type: ["string", "null"] }
        created_at:           { type: ["string", "null"] }
        updated_at:           { type: ["string", "null"] }
    LeadReview:
      type: object
      required: [review_status]
      properties:
        review_status:    { type: string }
        rejection_reason: { type: string }
    BulkLeadReview:
      type: object
      required: [lead_ids, review_status]
      properties:
        lead_ids:         { type: array, items: { type: string } }
        review_status:    { type: string }
        rejection_reason: { type: string }
    BulkReviewResult:
      type: object
      required: [updated_count]
      properties:
        updated_count: { type: integer }
...
    TriggerAnalysisInput:
      type: object
      required: [user_id, website_url, analysis_id]
      properties:
        user_id:     { type: string }
        website_url: { type: string }
        analysis_id: { type: string }
...
    WebhookResult:
      type: object
      required: [success]
      properties:
        success: { type: boolean }
        message: { type: ["string", "null"] }