# lib/api-zod/src/generated/types/outreachSettings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HubCredo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OutreachSettings {
  id: string;
  user_id: string;
  email_enabled?: boolean;
  linkedin_enabled?: boolean;
  /** @nullable */
  monthly_lead_target?: number | null;
  /** @nullable */
  monthly_email_target?: number | null;
  /** @nullable */
  monthly_linkedin_target?: number | null;
  /** @nullable */
  messaging_framework?: string | null;
  is_current?: boolean;
  /** @nullable */
  created_at?: string | null;
  /** @nullable */
  updated_at?: string | null;
}