# lib/api-zod/src/generated/types/stackInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HubCredo API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StackInput {
  icp_id: string;
  settings_id: string;
}