# lib/api-zod/src/generated/types/stackRecommendation.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HubCredo API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ToolItem } from './toolItem';

export interface StackRecommendation {
  id: string;
  user_id: string;
  /** @nullable */
  icp_id?: string | null;
  /** @nullable */
  settings_id?: string | null;
  tools: ToolItem[];
  is_current?: boolean;
  /** @nullable */
  created_at?: string | null;
}